requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-telegram-bot==21.6
pydantic==2.8.2
